Rename SauceConnectFourManager to SauceConnectManager
