+++ pom.xml
@@ -153,7 +153,7 @@
               <addDefaultSpecificationEntries>true</addDefaultSpecificationEntries>
               <addClasspath>true</addClasspath>
               <classpathPrefix>lib/</classpathPrefix>
-              <mainClass>com.saucelabs.ci.sauceconnect.SauceConnectFourManager</mainClass>
+              <mainClass>com.saucelabs.ci.sauceconnect.SauceConnectManager</mainClass>
             </manifest>
           </archive>
         </configuration>
